+++ tests/e2e/env/src/slack/reporter.js
@@ -64,23 +64,40 @@ const initializeSlack = () => {
  * @returns {Promise<void>}
  */
 export async function sendFailedTestMessageToSlack( testName ) {
-	const pr = initializeSlack();
-	if ( ! pr ) {
+	const { branch, commit, webUrl } = initializeSlack();
+	if ( ! branch ) {
 		return;
 	}
 	const web = initializeWeb();
 
 	try {
 		// Adding the app does not add the app user to the channel
-		await web.conversations.join( E2E_SLACK_CHANNEL );
+		await web.conversations.join({
+			channel: E2E_SLACK_CHANNEL,
+			token: E2E_SLACK_TOKEN,
+		});
+	} catch ( error ) {
+		// Check the code property and log the response
+		if ( error.code === ErrorCode.PlatformError || error.code === ErrorCode.RequestError ||
+			error.code === ErrorCode.RateLimitedError || error.code === ErrorCode.HTTPError ) {
+			console.log( error.data );
+		} else {
+			// Some other error, oh no!
+			console.log(
+				'Error joining channel',
+				error
+			);
+		}
+	}
+	try {
 		// For details, see: https://api.slack.com/methods/chat.postMessage
 		await web.chat.postMessage({
 			channel: E2E_SLACK_CHANNEL,
 			token: E2E_SLACK_TOKEN,
-			text: `Test failed on *${ pr.branch }* branch. \n
-            The commit this build is testing is *${ pr.commit }*. \n
+			text: `Test failed on *${ branch }* branch. \n
+            The commit this build is testing is *${ commit }*. \n
             The name of the test that failed: *${ testName }*. \n
-            See screenshot of the failed test below. *Build log* could be found here: ${ pr.webUrl }`,
+            See screenshot of the failed test below. *Build log* could be found here: ${ webUrl }`,
 		});
 	} catch ( error ) {
 		// Check the code property and log the response
@@ -91,7 +108,7 @@ export async function sendFailedTestMessageToSlack( testName ) {
 			// Some other error, oh no!
 			console.log(
 				'The error occurred does not match an error we are checking for in this block.',
-				error.data
+				error
 			);
 		}
 	}